Testing batch 0

+++ Code/multiprocess_lab.py
@@ -36,30 +36,32 @@ Grid = [{'Darwinian phi rate': {D_keys[i]: D_combination[i] for i in range(len(D
          'genetic parameters': {g_keys[i]: g_combination[i] for i in range(len(g_keys))}} for D_combination in D_combinations for e_combination in e_combinations for g_combination in g_combinations]
 
 env = E_CVRP_TW(path)
-# test_batch = env.generate_test_batch_per_size('l',2)
-# test_batch_num = 0
-# test_batch = test_batch[test_batch_num]
+test_batch = env.generate_test_batch_per_size('l',2)
+test_batch_num = 0
+test_batch = test_batch[test_batch_num]
 
-test_batch = env.generate_test_batch_per_size(['s','m'],1)
-test_batch_num = 's & m'
+# test_batch = env.generate_test_batch_per_size(['s','m'],1)
+# test_batch_num = 's & m'
 
 
 if __name__ == '__main__':
     for num, Configs in enumerate(Grid):
-        # with open(path + f'Experimentation/Exp {num}/readme.txt', 'w') as f:
-        #     readme = f'Experiment {num}'
-        #     readme += f'\nDarwinian phi rate: \t{Configs["Darwinian phi rate"]["penalization"]} - {Configs["Darwinian phi rate"]["length restriction"]}'
-        #     readme += f'\nevaluated insertion: \t{Configs["evaluated insertion"]["penalization"]} - {Configs["evaluated insertion"]["criterion"]}'
-        #     readme += f'\ngenetic configuration: \t{Configs["genetic parameters"]["population size"]} - {Configs["genetic parameters"]["crossover rate"]} - {Configs["genetic parameters"]["mutation rate"]}'
-        #     f.write(readme)
+        if num >= 92:
+            with open(path + f'Experimentation/Exp {num}/readme.txt', 'w') as f:
+                readme = f'Experiment {num}'
+                readme += f'\nDarwinian phi rate: \t{Configs["Darwinian phi rate"]["penalization"]} - {Configs["Darwinian phi rate"]["length restriction"]}'
+                readme += f'\nevaluated insertion: \t{Configs["evaluated insertion"]["penalization"]} - {Configs["evaluated insertion"]["criterion"]}'
+                readme += f'\ngenetic configuration: \t{Configs["genetic parameters"]["population size"]} - {Configs["genetic parameters"]["crossover rate"]} - {Configs["genetic parameters"]["mutation rate"]}'
+                f.write(readme)
+                    
 
-        progress_percentage = round(round((num+1)/len(Grid),4)*100,2)
-        print(f'\n-------- batch {test_batch_num} / Experiment {num} / {progress_percentage}% --------')
+            progress_percentage = round(round((num+1)/len(Grid),4)*100,2)
+            print(f'\n-------- batch {test_batch_num} / Experiment {num} / {progress_percentage}% --------')
 
-        lab:Experiment = Experiment(path, Operators, Configs, False, True, num)
+            lab:Experiment = Experiment(path, Operators, Configs, False, True, num)
 
-        if computer == 'mac':   p = pool.Pool(processes = 8)
-        else: p = pool.Pool()
+            if computer == 'mac':   p = pool.Pool(processes = 8)
+            else: p = pool.Pool()
 
-        p.map(lab.experimentation, test_batch)
-        p.terminate()
+            p.map(lab.experimentation, test_batch)
+            p.terminate()
